+++ contracts/multisig-prover/src/lib.rs
@@ -6,7 +6,6 @@ pub mod msg;
 mod payload;
 mod state;
 
-pub use encoding::Encoder;
 pub use payload::Payload;
 
 #[cfg(test)]
