+++ contracts/multisig-prover/src/msg.rs
@@ -4,58 +4,11 @@ use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{HexBinary, Uint64};
 use msgs_derive::EnsurePermissions;
 use multisig::key::KeyType;
+pub use multisig_prover_api::msg::InstantiateMsg;
 use router_api::CrossChainId;
 
 pub use crate::contract::MigrateMsg;
-use crate::{Encoder, Payload};
-
-#[cw_serde]
-pub struct InstantiateMsg {
-    /// Address that can execute all messages that either have unrestricted or admin permission level, such as Updateverifier set.
-    /// Should be set to a trusted address that can react to unexpected interruptions to the contract's operation.
-    pub admin_address: String,
-    /// Address that can call all messages of unrestricted, admin and governance permission level, such as UpdateSigningThreshold.
-    /// This address can execute messages that bypasses verification checks to rescue the contract if it got into an otherwise unrecoverable state due to external forces.
-    /// On mainnet, it should match the address of the Cosmos governance module.
-    pub governance_address: String,
-    /// Address of the gateway on axelar associated with the destination chain. For example, if this prover is creating proofs to
-    /// be relayed to Ethereum, this is the address of the gateway on Axelar for Ethereum.
-    pub gateway_address: String,
-    /// Address of the multisig contract on axelar.
-    pub multisig_address: String,
-    /// Address of the coordinator contract on axelar.
-    pub coordinator_address: String,
-    /// Address of the service registry contract on axelar.
-    pub service_registry_address: String,
-    /// Address of the voting verifier contract on axelar associated with the destination chain. For example, if this prover is creating
-    /// proofs to be relayed to Ethereum, this is the address of the voting verifier for Ethereum.
-    pub voting_verifier_address: String,
-    /// Threshold of weighted signatures required for signing to be considered complete
-    pub signing_threshold: MajorityThreshold,
-    /// Name of service in the service registry for which verifiers are registered.
-    pub service_name: String,
-    /// Name of chain for which this prover contract creates proofs.
-    pub chain_name: String,
-    /// Maximum tolerable difference between currently active verifier set and registered verifier set.
-    /// The verifier set registered in the service registry must be different by more than this number
-    /// of verifiers before calling UpdateVerifierSet. For example, if this is set to 1, UpdateVerifierSet
-    /// will fail unless the registered verifier set and active verifier set differ by more than 1.
-    pub verifier_set_diff_threshold: u32,
-    /// Type of encoding to use for signed payload. Blockchains can encode their execution payloads in various ways (ABI, BCS, etc).
-    /// This defines the specific encoding type to use for this prover, which should correspond to the encoding type used by the gateway
-    /// deployed on the destination chain.
-    pub encoder: Encoder,
-    /// Public key type verifiers use for signing payload. Different blockchains support different cryptographic signature algorithms (ECDSA, Ed25519, etc).
-    /// This defines the specific signature algorithm to use for this prover, which should correspond to the signature algorithm used by the gateway
-    /// deployed on the destination chain. The multisig contract supports multiple public keys per verifier (each a different type of key), and this
-    /// parameter controls which registered public key to use for signing for each verifier registered to the destination chain.
-    pub key_type: KeyType,
-    /// An opaque value created to distinguish distinct chains that the external gateway should be initialized with.
-    /// Value must be a String in hex format without `0x`, e.g. "598ba04d225cec385d1ce3cf3c9a076af803aa5c614bc0e0d176f04ac8d28f55".
-    #[serde(with = "axelar_wasm_std::hex")] // (de)serialization with hex module
-    #[schemars(with = "String")] // necessary attribute in conjunction with #[serde(with ...)]
-    pub domain_separator: Hash,
-}
+use crate::Payload;
 
 #[cw_serde]
 #[derive(EnsurePermissions)]
