+++ contracts/voting-verifier/Cargo.toml
@@ -46,6 +46,7 @@ serde_json = { workspace = true }
 service-registry = { workspace = true, features = ["library"] }
 service-registry-api = { workspace = true }
 thiserror = { workspace = true }
+voting-verifier-api = { workspace = true }
 
 [dev-dependencies]
 alloy-primitives = { version = "0.7.7", features = ["getrandom"] }
