+++ packages/multisig-prover-api/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "multisig-prover-api"
+version = "1.0.0"
+rust-version = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+axelar-wasm-std = { workspace = true, features = ["derive"] }
+cosmwasm-schema = { workspace = true }
+multisig = { workspace = true, features = ["library"] }
+
+[lints]
+workspace = true
