+++ packages/multisig-prover-api/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod encoding;
+pub mod msg;
